+++ importers/tests/Service/VendorService/DataWell/DataConverter/AmazonPublicUrlConverterTest.php
@@ -6,10 +6,10 @@
 
 namespace Tests\Service\VendorService\DataWell\DataConverter;
 
-use App\Service\VendorService\DataWell\DataConverter\IversePublicUrlConverter;
+use App\Service\VendorService\DataWell\DataConverter\AmazonPublicUrlConverter;
 use PHPUnit\Framework\TestCase;
 
-class IversePublicUrlConverterTest extends TestCase
+class AmazonPublicUrlConverterTest extends TestCase
 {
     /**
      * Test that array of urls are changed correctly.
@@ -29,7 +29,7 @@ class IversePublicUrlConverterTest extends TestCase
             '9788799996438' => 'https://s3.amazonaws.com/iverse_public/store/cover/padlarge/7775000045.png',
         ];
 
-        IversePublicUrlConverter::convertArrayValues($list);
+        AmazonPublicUrlConverter::convertArrayValues($list);
         $this->assertEquals($expected, $list);
 
         // Non 'Iverse' urls should remain unchanged
@@ -45,7 +45,7 @@ class IversePublicUrlConverterTest extends TestCase
             '9788799996438' => 'https://images.pubhub.dk/originals/ffe622ff-b267-40ea-a3ae-d1f2461075db.jpg',
         ];
 
-        IversePublicUrlConverter::convertArrayValues($list);
+        AmazonPublicUrlConverter::convertArrayValues($list);
         $this->assertEquals($expected, $list);
     }
 
@@ -56,24 +56,24 @@ class IversePublicUrlConverterTest extends TestCase
     {
         // Convert 'Iverse' urls correctly
         $original = 'https://s3.amazonaws.com/iverse_public/store/cover/padmedium/7775000045.png';
-        $converted = IversePublicUrlConverter::convertSingleUrl($original);
+        $converted = AmazonPublicUrlConverter::convertSingleUrl($original);
         $this->assertEquals('https://s3.amazonaws.com/iverse_public/store/cover/padlarge/7775000045.png', $converted);
 
         $original = 'https://s3.amazonaws.com/iverse_public/store/cover/padmedium/0000000792.jpg';
-        $converted = IversePublicUrlConverter::convertSingleUrl($original);
+        $converted = AmazonPublicUrlConverter::convertSingleUrl($original);
         $this->assertEquals('https://s3.amazonaws.com/iverse_public/store/cover/padlarge/0000000792.jpg', $converted);
 
         // Non 'Iverse' urls should remain unchanged
         $original = 'http://ebookcentral.proquest.com/covers/1752705-l.jpg';
-        $converted = IversePublicUrlConverter::convertSingleUrl($original);
+        $converted = AmazonPublicUrlConverter::convertSingleUrl($original);
         $this->assertEquals($original, $converted);
 
         $original = 'https://images.bogportalen.dk/images/9788799783205.jpg';
-        $converted = IversePublicUrlConverter::convertSingleUrl($original);
+        $converted = AmazonPublicUrlConverter::convertSingleUrl($original);
         $this->assertEquals($original, $converted);
 
         $original = 'https://imgcdn.saxo.com/_9788799997831/0x0';
-        $converted = IversePublicUrlConverter::convertSingleUrl($original);
+        $converted = AmazonPublicUrlConverter::convertSingleUrl($original);
         $this->assertEquals($original, $converted);
     }
 }